server/requirements.txt
fastapi
strawberry-graphql
psycopg2-binary
python-dotenv
asyncpg